[MOV] barcodes: split `barcodes.py`
Separates `barcode.nomenclature` and `barcode.rule` in their own file.

task-1968113
