+++ addons/barcodes/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
-from . import barcodes
 from . import barcode_events_mixin
+from . import barcode_nomenclature
+from . import barcode_rule
 from . import ir_http
 from . import res_company
